GREEN Successful compilation and package ready for docker

+++ src/main/java/com/techreturners/masterofarts/weatheradvisor/model/Weather.java
@@ -47,7 +47,7 @@ public class Weather {
      * Rainfall in last hour in mm
      */
     @Schema(description = "Rainfall in last hour in mm")
-    long rain;
+    double rain;
 
     /**
      * Current cloud cover as %
